Extract reusable dataset form
